# frontend/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "safehaven",
    "description": "",
    "license": {
      "name": ""
    },
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/admin/entities": {
      "post": {
        "tags": [
          "admin::entities"
        ],
        "operationId": "admin_entity_new",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/NewEntity"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Entity created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Entity"
                }
              }
            }
          },
          "401": {
            "description": "Invalid permissions",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/admin/entities/pending": {
      "get": {
        "tags": [
          "admin::entities"
        ],
        "operationId": "admin_entities_pending",
        "responses": {
          "200": {
            "description": "List of pending entities",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ListedEntity"
                  }
                }
              }
            }
          },
          "401": {
            "description": "Invalid permissions",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/admin/entities/search": {
      "get": {
        "tags": [
          "admin::entities"
        ],
        "operationId": "admin_entities_search",
        "parameters": [
          {
            "name": "search",
            "in": "query",
            "description": "Search query",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "page",
            "in": "query",
            "description": "Current page (default: 1)",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          },
          {
            "name": "page_size",
            "in": "query",
            "description": "Number of items per page (default: 20)",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Search results for entities",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ListedEntity"
                  }
                }
              }
            }
          },
          "401": {
            "description": "Invalid permissions",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/admin/entities/{id}": {
      "get": {
        "tags": [
          "admin::entities"
        ],
        "operationId": "admin_entity_get",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Entity identifier",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Entity details",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Entity"
                }
              }
            }
          },
          "401": {
            "description": "Invalid permissions",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "Not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "admin::entities"
        ],
        "operationId": "admin_entity_update",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Entity identifier",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateEntity"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Entity updated",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Entity"
                }
              }
            }
          },
          "401": {
            "description": "Invalid permissions",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "Not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "admin::entities"
        ],
        "operationId": "admin_entity_delete",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Entity identifier",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Entity deleted successfully"
          },
          "401": {
            "description": "Invalid permissions",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "Not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/admin/entities/{id}/comments": {
      "get": {
        "tags": [
          "admin::entities"
        ],
        "operationId": "admin_entity_get_comments",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Entity identifier",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "List of comments for the entity",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ListedComment"
                  }
                }
              }
            }
          },
          "401": {
            "description": "Invalid permissions",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/admin/entities/{parent_id}/parent/{child_id}": {
      "post": {
        "tags": [
          "admin::entities"
        ],
        "operationId": "admin_entity_register_parent",
        "parameters": [
          {
            "name": "parent_id",
            "in": "path",
            "description": "Parent entity identifier",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "child_id",
            "in": "path",
            "description": "Child entity identifier",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Parent entity registered successfully"
          },
          "401": {
            "description": "Invalid permissions",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "admin::entities"
        ],
        "operationId": "admin_entity_remove_parent",
        "parameters": [
          {
            "name": "parent_id",
            "in": "path",
            "description": "Parent entity identifier",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "child_id",
            "in": "path",
            "description": "Child entity identifier",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Parent-child relationship removed successfully"
          },
          "401": {
            "description": "Invalid permissions",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "Relationship not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "Entity": {
        "type": "object",
        "required": [
          "id",
          "display_name",
          "category_id",
          "locations",
          "data",
          "hide_from_map",
          "created_at",
          "updated_at"
        ],
        "properties": {
          "category_id": {
            "type": "string",
            "format": "uuid"
          },
          "created_at": {
            "type": "string",
            "format": "date-time"
          },
          "data": {},
          "display_name": {
            "type": "string"
          },
          "hide_from_map": {
            "type": "boolean"
          },
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "last_update_by": {
            "type": "string",
            "format": "uuid",
            "nullable": true
          },
          "locations": {
            "$ref": "#/components/schemas/Json"
          },
          "moderated_at": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "moderation_notes": {
            "type": "string",
            "nullable": true
          },
          "updated_at": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
      "ErrorResponse": {
        "type": "object",
        "required": [
          "error_code"
        ],
        "properties": {
          "details": {
            "type": "string",
            "nullable": true
          },
          "error_code": {
            "type": "string"
          }
        }
      },
...
      "ListedComment": {
        "type": "object",
        "required": [
          "id",
          "entity_id",
          "entity_display_name",
          "entity_category_id",
          "created_at"
        ],
        "properties": {
          "created_at": {
            "type": "string",
            "format": "date-time"
          },
          "entity_category_id": {
            "type": "string",
            "format": "uuid"
          },
          "entity_display_name": {
            "type": "string"
          },
          "entity_id": {
            "type": "string",
            "format": "uuid"
          },
          "id": {
            "type": "string",
            "format": "uuid"
          }
        }
      },
      "ListedEntity": {
        "type": "object",
        "required": [
          "id",
          "display_name",
          "category_id",
          "created_at"
        ],
        "properties": {
          "category_id": {
            "type": "string",
            "format": "uuid"
          },
          "created_at": {
            "type": "string",
            "format": "date-time"
          },
          "display_name": {
            "type": "string"
          },
          "id": {
            "type": "string",
            "format": "uuid"
          }
        }
      },
...
      "NewEntity": {
        "type": "object",
        "required": [
          "display_name",
          "category_id",
          "data",
          "last_update_by"
        ],
        "properties": {
          "category_id": {
            "type": "string",
            "format": "uuid"
          },
          "data": {},
          "display_name": {
            "type": "string"
          },
          "last_update_by": {
            "type": "string",
            "format": "uuid"
          }
        }
      },
...
      "UpdateEntity": {
        "type": "object",
        "required": [
          "display_name",
          "category_id",
          "data"
        ],
        "properties": {
          "category_id": {
            "type": "string",
            "format": "uuid"
          },
          "data": {},
          "display_name": {
            "type": "string"
          },
          "last_update_by": {
            "type": "string",
            "format": "uuid",
            "nullable": true
          },
          "moderation_notes": {
            "type": "string",
            "nullable": true
          }
        }
      },
...
    }
  }
}